requirements.txt
flask==3.0.3
flask-socketio==5.3.6
eventlet==0.35.2
google-generativeai==0.7.2
gTTS==2.5.1
faster-whisper==1.0.3
langdetect==1.0.9
soundfile==0.12.1
